Create a UIComponentList & RunApplication function
UIComponentList : Stores all the UIComponents of the application in the
order they should be drawn onto the screen.

RunApplication : A utility function to initialize the UIComponents of
the app and run the main loop (rendering and event handling loop).

Change the Makefile to compile a new source file `app.cpp`

Some comments were changed in the `display.cpp`