Add the ability to move inside the viewport
Added a Background UIComponent to represent the applications background.

Added `handleMouseEvent` virtual function to the base class UIComponent

Added `isPointInside` virtual function to check if a given point is inside a UIComponent or not.

Added a `activeComponent` attribute to the UIComponentList to keep track of the UIComponent on which the mouse pointer is hovering.

Defined a custom `handleMouseEvent` function for the Viewport class

Added a "hack" renderCover method to the viewport to stop objects from spilling out of the viewport

Changed the viewport's constructor signature.

- Cleaned up the main function and added all the app's logic to the RunApplication function.